SGEData Class

An S7 class that holds loaded SGE screen data before analysis.

Usage

SGEData(
  counts = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
attr(out, "row.names") <- row.names
         out
     }

    })(),
  metadata = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  annotation = NULL,
  complete_dataset = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
attr(out, "row.names") <- row.names
      
      out
     }
 })()
)

Arguments

counts

A data frame of raw count data with SEQUENCE and SAMPLE columns.

metadata

A data frame of sample metadata with condition information.

annotation

A data frame of variant annotations (optional, can be NULL).

complete_dataset

A data frame of counts joined with metadata.

Value

An S7 object of class SGEData.

Details

Access properties using the @ operator:

data@counts - Raw count data

data@metadata - Sample metadata

data@annotation - Variant annotations (may be NULL)

data@complete_dataset - Counts joined with metadata

Examples

if (FALSE) { # \dontrun{
# Load data (returns SGEData object)
data <- load_sge_data(
  count_files = "inputs/",
  metadata_file = "metadata/screen_metadata.tsv"
)

# Print shows summary
data

# Access properties
data@counts
data@metadata
data@annotation
data@complete_dataset
} # }